bootutil: (refactor) split image_validate.c into separate files
Move bootutil_find_key(), bootutil_img_hash(), and
bootutil_img_security_cnt() into their own source files for cleaner

No functional changes have been made.
